requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
certifi==2019.9.11
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
distlib==0.3.0
dj-database-url==0.5.0
Django==2.1.5
django-cors-headers==2.5.2
django-extensions==2.2.5
django-heroku==0.3.1
djangorestframework==3.8.2
filelock==3.0.12
flake8==3.8.2
gunicorn==19.9.0
identify==1.4.16
idna==2.8
importlib-metadata==1.6.0
mccabe==0.6.1
nodeenv==1.3.5
pathspec==0.8.0
pre-commit==2.4.0
pandas==1.2.1
parameterized==0.7.4
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycodestyle==2.6.0
pydotplus==2.0.2
pyflakes==2.2.0
pyparsing==2.4.5
pytz==2018.5
PyYAML==5.3.1
regex==2020.5.14
requests==2.22.0
six==1.13.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.6
virtualenv==20.0.21
whitenoise==4.1.4
zipp==3.1.0
